CRYPTOPP 5.6.3 RC6 checkin

+++ arc4.cpp
@@ -37,11 +37,11 @@ void ARC4_Base::UncheckedSetKey(const byte *key, unsigned int keyLen, const Name
 	unsigned int keyIndex = 0, stateIndex = 0;
 	for (i=0; i<256; i++)
 	{
-		byte a = m_state[i];
+		unsigned int a = m_state[i];
 		stateIndex += key[keyIndex] + a;
 		stateIndex &= 0xff;
 		m_state[i] = m_state[stateIndex];
-		m_state[stateIndex] = a;
+		m_state[stateIndex] = byte(a);
 		if (++keyIndex >= keyLen)
 			keyIndex = 0;
 	}
@@ -53,19 +53,19 @@ void ARC4_Base::UncheckedSetKey(const byte *key, unsigned int keyLen, const Name
 template <class T>
 static inline unsigned int MakeByte(T &x, T &y, byte *s)
 {
-	byte a = s[x];
-	y = (y+a) & 0xff;
-	byte b = s[y];
-	s[x] = b;
-	s[y] = a;
-	x = (x+1) & 0xff;
+	unsigned int a = s[x];
+	y = byte((y+a) & 0xff);
+	unsigned int b = s[y];
+	s[x] = byte(b);
+	s[y] = byte(a);
+	x = byte((x+1) & 0xff);
 	return s[(a+b) & 0xff];
 }
 
 void ARC4_Base::GenerateBlock(byte *output, size_t size)
 {
 	while (size--)
-		*output++ = (byte)MakeByte(m_x, m_y, m_state);
+		*output++ = static_cast<byte>(MakeByte(m_x, m_y, m_state));
 }
 
 void ARC4_Base::ProcessData(byte *outString, const byte *inString, size_t length)
@@ -88,7 +88,7 @@ void ARC4_Base::ProcessData(byte *outString, const byte *inString, size_t length
 	{
 		do
 		{
-			*outString++ = *inString++ ^ MakeByte(x, y, s);
+			*outString++ = *inString++ ^ byte(MakeByte(x, y, s));
 		}
 		while(--length);
 	}
